requirements.txt
livekit
livekit-agents
livekit-api
livekit-plugins-google
livekit-plugins-cartesia
livekit-plugins-groq
livekit-plugins-silero
pymongo
redis
python-dotenv